+++ modal/experimental.py
@@ -1,12 +1,22 @@
 # Copyright Modal Labs 2022
+import typing
 from typing import (
     Any,
     Callable,
+    List,
 )
 
+import modal
+
 from ._container_io_manager import _ContainerIOManager
-from .exception import InvalidError
-from .functions import _Function
+from ._utils.async_utils import (
+    synchronize_api,
+)
+from .exception import (
+    InvalidError,
+)
+from .functions import FunctionCall, OriginalReturnType, P, ReturnType, _Function
+from .object import _Object
 from .partial_function import _PartialFunction, _PartialFunctionFlags
 
 
@@ -24,6 +34,34 @@ def get_local_input_concurrency():
 
 
 # START Experimental: Container Networking
+
+
+class _GroupedFunction(typing.Generic[P, ReturnType, OriginalReturnType], _Object, type_prefix="gf"):
+    def __init__(self, f: _Function, size: int):
+        self.f = synchronize_api(f)
+        self.size = size
+
+    def remote(self, *args: P.args, **kwargs: P.kwargs) -> List[ReturnType]:
+        """
+        Calls the function remotely, executing it with the given arguments and returning the execution's result.
+        """
+        worker_handles: List[FunctionCall] = []
+        with modal.Queue.ephemeral() as q:
+            for i in range(self.size):
+                handle = self.f.spawn(*args, **{**kwargs, "rank": i, "size": self.size, "q": q})
+                worker_handles.append(handle)
+        output: list[ReturnType] = []
+        for i, handle in enumerate(worker_handles):
+            output.append(handle.get())
+        return output
+
+    def local(self, *args: P.args, **kwargs: P.kwargs) -> ReturnType:
+        raise NotImplementedError("Grouped function cannot be run locally")
+
+    def spawn(self, *args: P.args, **kwargs: P.kwargs) -> ReturnType:
+        raise NotImplementedError("Grouped function cannot be spawned")
+
+
 def grouped(size=1):
     def wrapper(raw_f: Callable[..., Any]) -> _PartialFunction:
         if isinstance(raw_f, _Function):
